% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepktrum.R
\name{list_spektrum_colors}
\alias{list_spektrum_colors}
\title{Function to list all spektrum colors with hex code ####}
\usage{
list_spektrum_colors(...)
}
\description{
Function to list all spektrum colors with hex code ####
}
