% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepktrum.R
\name{list_spektrum_palettes}
\alias{list_spektrum_palettes}
\title{Function to list all spektrum palettes####}
\usage{
list_spektrum_palettes(...)
}
\description{
Function to list all spektrum palettes####
}
\keyword{palette}
\keyword{spektrum}
