% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepktrum.R
\name{scale_fill_spektrum}
\alias{scale_fill_spektrum}
\title{Fill scale constructor for spektrum}
\usage{
scale_fill_spektrum(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\description{
Fill scale constructor for spektrum
}
