% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepktrum.R
\name{scale_color_spektrum}
\alias{scale_color_spektrum}
\title{Functions for Scales in ggplot2 ####
color scale constructor for spektrum}
\usage{
scale_color_spektrum(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\description{
Functions for Scales in ggplot2 ####
color scale constructor for spektrum
}
