% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepktrum.R
\name{spektrum_pal}
\alias{spektrum_pal}
\title{Palette function####
@param palette Character name of palette in spektrum_palettes
@param reverse Boolean indicating whether the palette should be reversed
@param ... Additional arguments to pass to colorRampPalette()}
\usage{
spektrum_pal(palette = "main", reverse = FALSE, ...)
}
\description{
Palette function####
@param palette Character name of palette in spektrum_palettes
@param reverse Boolean indicating whether the palette should be reversed
@param ... Additional arguments to pass to colorRampPalette()
}
